Added api_request_maker class
Moved custom functions to a new parent class called "api_request_maker"
